requirements_test.txt
# Testing Dependencies
# Additional packages needed for comprehensive testing

# Core testing framework
pytest>=7.0.0
pytest-django>=4.5.0
pytest-cov>=4.0.0

# Mock and test utilities
mock>=4.0.0
factory-boy>=3.2.0

# Performance testing
psutil>=5.9.0

# Code coverage
coverage>=7.0.0

# Test reporting
pytest-html>=3.1.0
pytest-xdist>=3.0.0

# Load testing
locust>=2.0.0

# Security testing
bandit>=1.7.0
safety>=2.3.0

# Code quality
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0

# Documentation testing
sphinx>=6.0.0
sphinx-rtd-theme>=1.2.0
